Fix attendee reports crashing on empty list
When there were no stored reports available at all, both the load and
delete buttons would cause a crash. Fix that by both adding defense at
the server side returning a 404 instead of a crash in this case, and by
a client side check to not even try it in this case.

+++ postgresqleu/confreg/views.py
@@ -2682,6 +2682,9 @@ def reports(request, confname):
 
     if request.method == 'POST':
         if request.POST['what'] == 'delete':
+            if request.POST.get('storedreport', '') == '':
+                raise Http404()
+
             get_object_or_404(SavedReportDefinition, conference=conference, pk=request.POST['storedreport']).delete()
             return HttpResponse("OK")
 
@@ -2705,6 +2708,9 @@ def reports(request, confname):
         raise Http404()
     elif 'storedreport' in request.GET:
         # Load a special stored report
+        if request.GET.get('storedreport', '') == '':
+            raise Http404()
+
         r = get_object_or_404(SavedReportDefinition, conference=conference, pk=request.GET['storedreport'])
         return HttpResponse(json.dumps(r.definition), content_type='application/json')
